Refactor pub-sub logic some more
